Organize Core files
