+++ Sources/Core/Extensions/Collection.swift
