+++ Sources/Core/Types/DemandBuffer.swift
