+++ Sources/Core/Types/LimitedBuffer.swift
