+++ Sources/Core/Types/Trigger.swift
